combined structure and output tests into 1 end point and fixed tests
