+++ executeChallenges/src/app.ts
@@ -7,7 +7,6 @@ import { executeJSRouter } from './routes/executeChallengeJS'
 import { compileCRouter } from './routes/compileC'
 import { compileJSRouter } from './routes/compileJS'
 import { compileJavaRouter } from './routes/compileJava'
-import { detectJavaDesignPatternsRouter } from './routes/detectJavaDesignPatterns'
 
 const app = express()
 
@@ -25,7 +24,6 @@ app.use(executeJSRouter);
 app.use(compileCRouter);
 app.use(compileJSRouter);
 app.use(compileJavaRouter);
-app.use(detectJavaDesignPatternsRouter);
 
 app.use(ErrorHandler);
