+++ executeChallenges/src/routes/__test__/executeChallengeJava.test.ts
@@ -26,7 +26,7 @@ it('successfully runs tests', async()=>{
             ]
         }),
         language: "java",
-        expectedStructure: '',
+        expectedStructure: '[{"className":"Furniture","modifiers":[\"abstract interface\"],"superClass":"","interfaces":[],"constructors":[],"methods":[],"fields":[]},{"className":"furnitureFactory","modifiers":[\"abstract interface\"],"superClass":"","interfaces":[],"constructors":[],"methods":[{"modifiers":[\"public\", \"abstract\"],"name":"getFurniture","returnType":"Furniture","parameters":[]}],"fields":[]},{"className":"Chair","modifiers":[],"superClass":"","interfaces":["Furniture"],"constructors":[{"modifiers":[],"parameters":[]}],"methods":[],"fields":[]}]',
         expectedDesignPatterns: []
     })
     await challenge.save()
@@ -37,6 +37,7 @@ it('successfully runs tests', async()=>{
             solution: JSON.stringify(`class Solution { public int solution(int a,int b,int c){return(a+b+c);}}`)
         })
         .expect(200)
+    expect(result.body.data.structure).toEqual(false)
     expect(result.body.data.successfulTests).toEqual(result.body.data.totalTestsDone)
 
 })
@@ -65,7 +66,7 @@ it('successfully runs tests 2', async()=>{
         }),
         language: "java",
         expectedStructure: '',
-        expectedDesignPatterns: []
+        expectedDesignPatterns: ['factory']
     })
     await challenge.save()
     const result = await request(app)
@@ -75,6 +76,7 @@ it('successfully runs tests 2', async()=>{
             solution: JSON.stringify(`class Solution { public int solution(int a,int b,int c){return(a+b+c);}}`)
         })
         .expect(200)
+    expect(result.body.data.designPatterns.factory).toEqual(false)
     expect(result.body.data.successfulTests).toEqual(result.body.data.totalTestsDone-1)
 
 })
@@ -297,7 +299,7 @@ it('successfully hides arguments from errors', async()=>{
             ]
         }),
         expectedStructure: '',
-        expectedDesignPatterns: [],
+        expectedDesignPatterns: ['factory'],
         language: "java"
     })
     await challenge.save()
@@ -313,6 +315,59 @@ it('successfully hides arguments from errors', async()=>{
 
 })
 
+it('successfully detects all class names and interfaces', async()=>{
+    const user = new mongoose.Types.ObjectId()
+    const challenge = new Challenge({
+        status: 'approved',
+        startsAt: Date.now(),
+        expiresAt: "2014-02-01T00:00:00",
+        expectedOutputTests: '',
+        language: "java",
+        expectedStructure: '[{"className":"mlkia$peops","modifiers":[],"superClass":"","interfaces":[],"constructors":[],"methods":[{"modifiers":[\"private\"],"name":"getPepe","returnType":"int","parameters":[]}],"fields":[{"modifiers":[\"private\"],"name":"pepe","type":"int"}]},{"className":"mlkia","modifiers":[],"superClass":"","interfaces":[],"constructors":[],"methods":[],"fields":[]}]',
+        expectedDesignPatterns: []
+    })
+    await challenge.save()
+    const response = await request(app)
+        .post(`/api/v1/compile/challengejava/${challenge.id}`)
+        .set('Cookie', global.signin(user,'user'))
+        .send({
+            solution: JSON.stringify(`class mlkia {
+                class peops{
+                    private int pepe=5;
+                    private int getPepe(){
+                        return pepe;
+                    }
+                }
+            }
+            
+            abstract class mlkia2{
+            
+            }
+            
+            class mlkia3 implements pipis2{
+            
+            }
+            
+            class mlkia4 extends mlkia2{
+            
+            }
+            
+            interface pipis1 {
+            
+            }
+            
+            interface pipis extends pipis2{
+            
+            }
+            
+            interface pipis2{
+            
+            }`)
+        })
+        .expect(200)
+    expect(response.body.data.structure).toEqual(true)
+})
+
 it('successfully detects all class names and interfaces 2', async()=>{
     const user = new mongoose.Types.ObjectId()
     const challenge = new Challenge({
@@ -322,7 +377,7 @@ it('successfully detects all class names and interfaces 2', async()=>{
         expectedOutputTests: '',
         language: "java",
         expectedStructure: '[{"className":"TestEntity2","modifiers":[],"superClass":"TestEntitySuper","interfaces":["TestInt"],"constructors":[{"modifiers":[\"public\"],"parameters":[\"String\",\"int[]\",\"Map<String,Object>\"]}],"methods":[{"modifiers":[\"public\"],"name":"getM","returnType":"Map<String, Object>","parameters":[]},{"modifiers":[\"public\", \"static\"],"name":"testMethod","returnType":"void","parameters":[\"int\",\"String\",\"Integer\"]}],"fields":[{"modifiers":[\"private\"],"name":"m","type":"Map<String, Object>"},{"modifiers":[\"private static\"],"name":"peops","type":"TestEntity2"}]}]',
-        expectedDesignPatterns: ['singleton', 'factory']
+        expectedDesignPatterns: []
     })
     await challenge.save()
     const response = await request(app)
@@ -403,6 +458,285 @@ it('successfully detects all class names and interfaces 2', async()=>{
         })
         .expect(200)
     expect(response.body.data.structure).toEqual(true)
-    expect(response.body.data.designPatterns.singleton).toEqual(false)
+})
+
+it('successfully detects singleton', async()=>{
+    const user = new mongoose.Types.ObjectId()
+    const challenge = new Challenge({
+        status: 'approved',
+        startsAt: Date.now(),
+        expiresAt: "2014-02-01T00:00:00",
+        expectedOutputTests: '',
+        language: "java",
+        expectedStructure: '',
+        expectedDesignPatterns: ['singleton', 'factory', 'observer']
+    })
+    await challenge.save()
+    const response = await request(app)
+        .post(`/api/v1/compile/challengejava/${challenge.id}`)
+        .set('Cookie', global.signin(user,'user'))
+        .send({
+            solution: JSON.stringify(`class mlkia {
+                class peops{
+                    private int pepe=5;
+                    private int getPepe(){
+                        return pepe;
+                    }
+                }
+            }
+            
+            abstract class mlkia2{
+                
+            }
+
+            class Singleton{
+                private static Singleton sngl;
+                private Singleton(){
+
+                }
+                public static Singleton getSngl(){
+                    return sngl;
+                }
+            }
+            
+            class mlkia3 implements pipis2{
+            
+            }
+            
+            class mlkia4 extends mlkia2{
+            
+            }
+            
+            interface pipis1 {
+            
+            }
+            
+            interface pipis extends pipis2{
+            
+            }
+            
+            interface pipis2{
+            
+            }`)
+        })
+        .expect(200)
+    expect(response.body.data.designPatterns.singleton).toEqual(true)
     expect(response.body.data.designPatterns.factory).toEqual(false)
+    expect(response.body.data.designPatterns.observer).toEqual(false)
+})
+
+it('successfully detects factory', async()=>{
+    const user = new mongoose.Types.ObjectId()
+    const challenge = new Challenge({
+        status: 'approved',
+        startsAt: Date.now(),
+        expiresAt: "2014-02-01T00:00:00",
+        expectedOutputTests: '',
+        language: "java",
+        expectedStructure: '',
+        expectedDesignPatterns: ['singleton', 'factory', 'observer']
+    })
+    await challenge.save()
+    const response = await request(app)
+        .post(`/api/v1/compile/challengejava/${challenge.id}`)
+        .set('Cookie', global.signin(user,'user'))
+        .send({
+            solution: JSON.stringify(`
+            class mlkia {
+                class peops{
+                    private int pepe=5;
+                    private int getPepe(){
+                        return pepe;
+                    }
+                }
+            }
+            
+            abstract class Furniture{
+                
+            }
+
+            class Chair extends Furniture{
+
+            }
+            
+            interface furnitureFactory{
+                public Furniture getFurniture();
+            }
+            
+            class factorySubClass implements furnitureFactory{
+                public Furniture getFurniture(){
+
+                    return new Chair();
+                }
+            }
+            
+            `)
+        })
+        .expect(200)
+    expect(response.body.data.designPatterns.factory).toEqual(true)
+    expect(response.body.data.designPatterns.singleton).toEqual(false)
+    expect(response.body.data.designPatterns.observer).toEqual(false)
+})
+
+it('successfully detects factory 2', async()=>{
+    const user = new mongoose.Types.ObjectId()
+    const challenge = new Challenge({
+        status: 'approved',
+        startsAt: Date.now(),
+        expiresAt: "2014-02-01T00:00:00",
+        expectedOutputTests: '',
+        language: "java",
+        expectedStructure: '[{"className":"Furniture","modifiers":[\"abstract interface\"],"superClass":"","interfaces":[],"constructors":[],"methods":[],"fields":[]},{"className":"furnitureFactory","modifiers":[\"abstract interface\"],"superClass":"","interfaces":[],"constructors":[],"methods":[{"modifiers":[\"public\", \"abstract\"],"name":"getFurniture","returnType":"Furniture","parameters":[]}],"fields":[]},{"className":"Chair","modifiers":[],"superClass":"","interfaces":["Furniture"],"constructors":[{"modifiers":[],"parameters":[]}],"methods":[],"fields":[]}]',
+        expectedDesignPatterns: ['factory', 'singleton']
+    })
+    await challenge.save()
+    const response = await request(app)
+        .post(`/api/v1/compile/challengejava/${challenge.id}`)
+        .set('Cookie', global.signin(user,'user'))
+        .send({
+            solution: JSON.stringify(`
+            class mlkia {
+                class peops{
+                    private int pepe=5;
+                    private int getPepe(){
+                        return pepe;
+                    }
+                }
+            }
+            
+            interface Furniture{
+                
+            }
+
+            class Chair implements Furniture{
+
+            }
+            
+            interface furnitureFactory{
+                public Furniture getFurniture();
+            }
+            
+            class factorySubClass implements furnitureFactory{
+                public Furniture getFurniture(){
+
+                    return new Chair();
+                }
+            }
+            
+            `)
+        })
+        .expect(200)
+    expect(response.body.data.structure).toEqual(true)
+    expect(response.body.data.designPatterns.factory).toEqual(true)
+    expect(response.body.data.designPatterns.singleton).toEqual(false)
+    })
+
+it('successfully detects factory and singleton', async()=>{
+    const user = new mongoose.Types.ObjectId()
+    const challenge = new Challenge({
+        status: 'approved',
+        startsAt: Date.now(),
+        expiresAt: "2014-02-01T00:00:00",
+        expectedOutputTests: '',
+        language: "java",
+        expectedStructure: '[{"className":"Singleton","modifiers":[],"superClass":"","interfaces":[],"constructors":[{"modifiers":[\"private\"],"parameters":[]}],"methods":[{"modifiers":[\"public\", \"static\"],"name":"getSngl","returnType":"Singleton","parameters":[]}],"fields":[{"modifiers":[\"private static\"],"name":"sngl","type":"Singleton"}]}]',
+        expectedDesignPatterns: ['factory', 'singleton']
+    })
+    await challenge.save()
+    const response = await request(app)
+        .post(`/api/v1/compile/challengejava/${challenge.id}`)
+        .set('Cookie', global.signin(user,'user'))
+        .send({
+            solution: JSON.stringify(`
+            class mlkia {
+                class peops{
+                    private int pepe=5;
+                    private int getPepe(){
+                        return pepe;
+                    }
+                }
+            }
+            
+            interface Furniture{
+                
+            }
+
+            class Chair implements Furniture{
+
+            }
+            
+            interface furnitureFactory{
+                public Furniture getFurniture();
+            }
+            
+            class Singleton{
+                private static Singleton sngl;
+                private Singleton(){
+
+                }
+                public static Singleton getSngl(){
+                    return sngl;
+                }
+            }
+
+            class factorySubClass implements furnitureFactory{
+                public Furniture getFurniture(){
+
+                    return new Chair();
+                }
+            }
+            
+            `)
+        })
+        .expect(200)
+    expect(response.body.data.structure).toEqual(true)
+    expect(response.body.data.designPatterns.factory).toEqual(true)
+    expect(response.body.data.designPatterns.singleton).toEqual(true)
+})
+
+it('successfully detects factory and singleton', async()=>{
+    const user = new mongoose.Types.ObjectId()
+    const challenge = new Challenge({
+        status: 'approved',
+        startsAt: Date.now(),
+        expiresAt: "2014-02-01T00:00:00",
+        expectedOutputTests: '',
+        language: "java",
+        expectedStructure: '[{"className":"TestEntity2","modifiers":[],"superClass":"TestEntitySuper","interfaces":["TestInt"],"constructors":[{"modifiers":[\"public\"],"parameters":[\"String\",\"int[]\",\"Map<String, Object>\"]}],"methods":[{"modifiers":[\"public\"],"name":"getM","returnType":"Map<String, Object>","parameters":[]},{"modifiers":[\"public\", \"static\"],"name":"testMethod","returnType":"void","parameters":[\"int\",\"String\",\"Integer\"]}],"fields":[{"modifiers":[\"private\"],"name":"m","type":"Map<String, Object>"},{"modifiers":[\"private static\"],"name":"peops","type":"TestEntity2"}]}]',
+        expectedDesignPatterns: ['singleton', 'factory', 'observer']
+    })
+    await challenge.save()
+    const response = await request(app)
+        .post(`/api/v1/compile/challengejava/${challenge.id}`)
+        .set('Cookie', global.signin(user,'user'))
+        .send({
+            solution: JSON.stringify(`
+            class mlkia {
+                class peops{
+                    private int pepe=5;
+                    private int getPepe(){
+                        return pepe;
+                    }
+                }
+            }
+            
+            interface Observer{
+            }
+
+            class Chair implements Observer{
+                private Observable obs;
+            }
+            
+            class Observable{
+                private List<Observer> ls;
+            }
+            
+            
+            `)
+        })
+        .expect(200)
+    expect(response.body.data.structure).toEqual(false)
+    expect(response.body.data.designPatterns.factory).toEqual(false)
+    expect(response.body.data.designPatterns.singleton).toEqual(false)
+    expect(response.body.data.designPatterns.observer).toEqual(true)
 })
